TranslationWordColumn moved to controller/topic package
Added TranslationsListView
