+++ src/main/java/egl/client/controller/task/category/CategoryTheoryTaskController.java
@@ -1,9 +1,8 @@
 package egl.client.controller.task.category;
 
 import egl.client.controller.task.TheoryTaskController;
+import egl.client.controller.topic.TranslationsListView;
 import egl.client.model.topic.category.Category;
-import egl.client.model.topic.category.Translation;
-import egl.client.view.table.list.CustomListView;
 import javafx.fxml.FXML;
 import net.rgielen.fxweaver.core.FxmlView;
 import org.springframework.stereotype.Component;
@@ -15,7 +14,7 @@ import java.util.ResourceBundle;
 @FxmlView
 public class CategoryTheoryTaskController extends TheoryTaskController {
 
-    @FXML private CustomListView<Translation> translationsTableView;
+    @FXML private TranslationsListView translationsListView;
 
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
@@ -25,13 +24,13 @@ public class CategoryTheoryTaskController extends TheoryTaskController {
     @Override
     public void setPrefSize(double parentWidth, double parentHeight) {
         descriptionTextArea.setPrefSize(parentWidth, parentHeight * 0.3);
-        translationsTableView.setPrefSize(parentWidth, parentHeight * 0.6);
+        translationsListView.setPrefSize(parentWidth, parentHeight * 0.6);
     }
 
     @Override
     protected void prepareSpecificTheory() {
         Category category = (Category) controllerTopic;
         var categoryTranslations = category.getTranslations();
-        translationsTableView.setItems(categoryTranslations);
+        translationsListView.setItems(categoryTranslations);
     }
 }
